requirements.txt
aiofiles==23.2.1
aiohttp==3.8.6
autopep8==2.0.4
datedelta==1.2
dotmap==1.3.30
python-dotenv==0.19.2
pytz==2018.9
requests==2.21.0
stripe==2.64.0
